==> 00_Data_extraction.ipynb <==
import zipfile
import os

#Initial paths
zip_path = "~/work/MNLP-project-2/ocr_datasets.zip"
extract_dir = "~/work/MNLP-project-2/extracted"

# convert into absolute path
zip_path = os.path.expanduser(zip_path)
extract_dir = os.path.expanduser(extract_dir)

# Create the destination file if it doesn't already exist
os.makedirs(extract_dir, exist_ok=True)

# Opening and extracting the zip file
with zipfile.ZipFile(zip_path, 'r') as zip_ref:
    zip_ref.extractall(extract_dir)
    print("Fichiers extraits dans :", extract_dir)

# List the content
print("Contenu du ZIP :")
for file_name in zip_ref.namelist():
    print(file_name)
