==> 01_Collect_data.ipynb <==
# Download data

from google.colab import files
uploaded = files.upload()

import json

with open("the_vampyre_clean.json", "r", encoding="utf-8") as f:
    clean_data = json.load(f)

with open("the_vampyre_ocr.json", "r", encoding="utf-8") as f:
    ocr_data = json.load(f)

# Explore data

*   ocr_data

One text has been OCRed 48 times. Each key in the dictionnary ocr_data corresponds to one of this times.

*   clean_data
The clean texted, divided into 48 keys : each key corresponds to a part of the text. Key 0 is the begining of the text, while key 47 is the last element of the text. I suppose that the full correct text correspond to the concatenation of the 47 elements, with '\n' that separete them.



print("Type de clean_data :", type(clean_data))
print("Type de ocr_data   :", type(ocr_data))

print("Look at the_vampyre_clean.json :")
print("Avalaible keys :", list(clean_data.keys()))

print("Look at the_vampyre_ocr.json :")
print("Avalaible keys :", list(ocr_data.keys()))

print(clean_data['19'])

print(ocr_data['9'])

def concat_values_dict(d):
    """
    Concat values of a dict, seperating each element with '\n'

    Args:
        d (dict): Dictionnary

    Returns:
        str: concatenated text
    """
    return '\n'.join(d.get(str(i), "") for i in range(48))

clean_data_text = concat_values_dict(clean_data)
print(clean_data_text)
