==> 06_output_files.ipynb <==
import pandas as pd
import json

# "Clean" text

df_ocr = pd.read_csv('~/work/MNLP-project-2/data/eng/the_vampyre_ocr.csv')
df_t5 = pd.read_csv('T5_correction.csv')
df_bt = pd.read_csv('back_translation_correction.csv')
df_bart = pd.read_csv('bart_correction.csv')
df_bt_t5 = pd.read_csv('back_translation_t5_correction.csv')

df = df_ocr.copy()
model_name = 'ocr'

data_dict = dict(zip(df['id'], df['texte']))

with open(f"Pythoms-hw2_ocr-{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(data_dict, f, ensure_ascii=False, indent=4)

df = df_bt.copy()
model_name = 'back_translation'

data_dict = dict(zip(df['index'], df['back_translated_text']))

with open(f"Pythoms-hw2_ocr-{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(data_dict, f, ensure_ascii=False, indent=4)

df = df_bart.copy()
model_name = 'bart'

data_dict = dict(zip(df['index'], df['corrected_text']))

with open(f"Pythoms-hw2_ocr-{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(data_dict, f, ensure_ascii=False, indent=4)

df = df_t5.copy()
model_name = 't5'

data_dict = dict(zip(df['index'], df['corrected_text']))

with open(f"Pythoms-hw2_ocr-{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(data_dict, f, ensure_ascii=False, indent=4)

df = df_bt_t5.copy()
model_name = 'back_translation_t5'

data_dict = dict(zip(df['index'], df['corrected_text']))

with open(f"Pythoms-hw2_ocr-{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(data_dict, f, ensure_ascii=False, indent=4)

# Output of LLM as a judge

df_ocr = pd.read_csv('ocr_gemini_scores.csv')
df_t5 = pd.read_csv('T5_gemini_scores_all_texts.csv')
df_bt = pd.read_csv('Back_translation_gemini_scores_all_texts.csv')
df_bart = pd.read_csv('Bart_gemini_scores_all_texts.csv')
df_bt_t5 = pd.read_csv('back_translation_t5_gemini_scores_all_texts.csv')

df = df_ocr.copy()
model_name = 'ocr'

result = {}

for text_id, group in df.groupby('text_index'):
    chunks = group[['chunk_index', 'score']].to_dict(orient='records')
    result[str(text_id)] = chunks

with open(f"Pythoms-hw2_ocr-judge_{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(result, f, ensure_ascii=False, indent=4)

df = df_t5.copy()
model_name = 't5'

result = {}

for text_id, group in df.groupby('text_index'):
    chunks = group[['chunk_index', 'score']].to_dict(orient='records')
    result[str(text_id)] = chunks

with open(f"Pythoms-hw2_ocr-judge_{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(result, f, ensure_ascii=False, indent=4)

df = df_bt.copy()
model_name = 'back_translation'

result = {}

for text_id, group in df.groupby('text_index'):
    chunks = group[['chunk_index', 'score']].to_dict(orient='records')
    result[str(text_id)] = chunks

with open(f"Pythoms-hw2_ocr-judge_{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(result, f, ensure_ascii=False, indent=4)

df = df_bart.copy()
model_name = 'bart'

result = {}

for text_id, group in df.groupby('text_index'):
    chunks = group[['chunk_index', 'score']].to_dict(orient='records')
    result[str(text_id)] = chunks

with open(f"Pythoms-hw2_ocr-judge_{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(result, f, ensure_ascii=False, indent=4)

df = df_bt_t5.copy()
model_name = 'back_translation_t5'

result = {}

for text_id, group in df.groupby('text_index'):
    chunks = group[['chunk_index', 'score']].to_dict(orient='records')
    result[str(text_id)] = chunks

with open(f"Pythoms-hw2_ocr-judge_{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(result, f, ensure_ascii=False, indent=4)

# Manual annotations

df_ocr = pd.read_csv('ocr_manual_scores.csv')

df_t5 = pd.read_csv('T5_manual_scores.csv')

df_bt = pd.read_csv('Back_translation_manual_scores.csv')

df_bart = pd.read_csv('bart_manual_scores.csv')

df_bt_t5 = pd.read_csv('Back_translation_t5_manual_scores.csv')

df = df_ocr.copy()
model_name = 'ocr'

result = {}

for _, row in df.iterrows():
    text_id = str(row['text_index'])

    chunks = [
        {"type": "fixed", "text": row['fixed_chunk'], "score": row['score_fixed']},
        {"type": "random_1", "text": row['random_chunk_1'], "score": row['score_random_1']},
        {"type": "random_2", "text": row['random_chunk_2'], "score": row['score_random_2']}
    ]

    result[text_id] = chunks

with open(f"Pythoms-hw2_ocr-manual_{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(result, f, ensure_ascii=False, indent=4)

df = df_t5.copy()
model_name = 't5'

result = {}

for _, row in df.iterrows():
    text_id = str(row['text_index'])

    chunks = [
        {"type": "fixed", "text": row['fixed_chunk'], "score": row['score_fixed']},
        {"type": "random_1", "text": row['random_chunk_1'], "score": row['score_random_1']},
        {"type": "random_2", "text": row['random_chunk_2'], "score": row['score_random_2']}
    ]

    result[text_id] = chunks

with open(f"Pythoms-hw2_ocr-manual_{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(result, f, ensure_ascii=False, indent=4)

df = df_bt.copy()
model_name = 'back_translation'

result = {}

for _, row in df.iterrows():
    text_id = str(row['text_index'])

    chunks = [
        {"type": "fixed", "text": row['fixed_chunk'], "score": row['score_fixed']},
        {"type": "random_1", "text": row['random_chunk_1'], "score": row['score_random_1']},
        {"type": "random_2", "text": row['random_chunk_2'], "score": row['score_random_2']}
    ]

    result[text_id] = chunks

with open(f"Pythoms-hw2_ocr-manual_{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(result, f, ensure_ascii=False, indent=4)

df = df_bart.copy()
model_name = 'bart'

result = {}

for _, row in df.iterrows():
    text_id = str(row['text_index'])

    chunks = [
        {"type": "fixed", "text": row['fixed_chunk'], "score": row['score_fixed']},
        {"type": "random_1", "text": row['random_chunk_1'], "score": row['score_random_1']},
        {"type": "random_2", "text": row['random_chunk_2'], "score": row['score_random_2']}
    ]

    result[text_id] = chunks

with open(f"Pythoms-hw2_ocr-manual_{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(result, f, ensure_ascii=False, indent=4)

df = df_bt_t5.copy()
model_name = 'back_translation_t5'

result = {}

for _, row in df.iterrows():
    text_id = str(row['text_index'])

    chunks = [
        {"type": "fixed", "text": row['fixed_chunk'], "score": row['score_fixed']},
        {"type": "random_1", "text": row['random_chunk_1'], "score": row['score_random_1']},
        {"type": "random_2", "text": row['random_chunk_2'], "score": row['score_random_2']}
    ]

    result[text_id] = chunks

with open(f"Pythoms-hw2_ocr-manual_{model_name}.json", "w", encoding="utf-8") as f:
    json.dump(result, f, ensure_ascii=False, indent=4)

# First 5000 characters

bart_text = pd.read_csv("bart_correction.csv")
bart_text.head()

lines = bart_text['corrected_text'][0].split('\n')[:24]

indexed_lines = {i: line for i, line in enumerate(lines)}

with open('Pythoms-hw2_ocr_first_5k_tokens.json', 'w', encoding='utf-8') as f:
    json.dump(indexed_lines, f, ensure_ascii=False, indent=4)
